+++ Jelly/Classes/JellyFadeInPresentation.swift
@@ -5,7 +5,7 @@
 
 import Foundation
 
-public struct JellyFadeInPresentation: JellyPresentation {
+public struct JellyFadeInPresentation: JellyPresentation, AlignablePresentation {
     
     // Jelly Presentation Protocol conformance
     public private(set) var dismissCurve: JellyConstants.JellyCurve = .linear
@@ -13,21 +13,31 @@ public struct JellyFadeInPresentation: JellyPresentation {
     public private(set) var cornerRadius: Double = 0.0
     public private(set) var backgroundStyle: JellyConstants.BackgroundStyle = .dimmed
     public private(set) var duration : JellyConstants.Duration = .normal // Duration the ViewController needs to kick in
-    public private(set) var sizeForViewController: CGSize = CGSize(width: 300, height: 300) // Size for the presented ViewController
+    public private(set) var widthForViewController: JellyConstants.Size = .fullscreen
+    public private(set) var heightForViewController: JellyConstants.Size = .fullscreen
+    public private(set) var marginGuards: UIEdgeInsets = .zero
+
+    // Alginable
+    public private(set) var horizontalAlignment: JellyConstants.HorizontalAlignment = .center
+    public private(set) var verticalAlignemt: JellyConstants.VerticalAlignment = .center
     
     public init(dismissCurve: JellyConstants.JellyCurve = .linear,
                 presentationCurve: JellyConstants.JellyCurve = .linear,
                 cornerRadius: Double = 0.0,
                 backgroundStyle: JellyConstants.BackgroundStyle = .dimmed,
                 duration: JellyConstants.Duration = .normal,
-                sizeForViewController: CGSize = CGSize(width:300,height:300)) {
+                widthForViewController: JellyConstants.Size = .fullscreen,
+                heightForViewController: JellyConstants.Size = .fullscreen,
+                marginGuards: UIEdgeInsets = .zero) {
         
         self.dismissCurve = dismissCurve
         self.presentationCurve = presentationCurve
         self.cornerRadius = cornerRadius
         self.backgroundStyle = backgroundStyle
         self.duration = duration
-        self.sizeForViewController = sizeForViewController
+        self.widthForViewController = widthForViewController
+        self.heightForViewController = heightForViewController
+        self.marginGuards = marginGuards
     }
     
 }