+++ Jelly/Classes/JellySlideInPresentation.swift
@@ -3,7 +3,7 @@
 //  Created by Sebastian Boldt on 20.11.16.
 //
 
-public struct JellySlideInPresentation: JellyPresentation {
+public struct JellySlideInPresentation: JellyPresentation, AlignablePresentation {
     
     // JellyPresentation Protocol conformance
     public private(set) var dismissCurve: JellyConstants.JellyCurve = .linear
@@ -12,11 +12,17 @@ public struct JellySlideInPresentation: JellyPresentation {
     public private(set) var backgroundStyle: JellyConstants.BackgroundStyle = .none
     public private(set) var jellyness: JellyConstants.Jellyness
     public private(set) var duration : JellyConstants.Duration = .normal // Duration the ViewController needs to kick in
-    public private(set) var sizeForViewController: CGSize = CGSize(width: 300, height: 600) // Size for the presented ViewController
+    public private(set) var widthForViewController: JellyConstants.Size = .fullscreen
+    public private(set) var heightForViewController: JellyConstants.Size = .fullscreen
+    public private(set) var marginGuards: UIEdgeInsets = .zero
     
     // Unique
     public private(set) var directionShow: JellyConstants.Direction = .left // Direction the ViewController slides in from
     public private(set) var directionDismiss: JellyConstants.Direction = .left // Direction the ViewController slides out to
+    
+    // Alginable
+    public private(set) var horizontalAlignment: JellyConstants.HorizontalAlignment = .center
+    public private(set) var verticalAlignemt: JellyConstants.VerticalAlignment = .center
     
     public init(dismissCurve: JellyConstants.JellyCurve = .linear,
                 presentationCurve: JellyConstants.JellyCurve = .linear,
@@ -24,9 +30,13 @@ public struct JellySlideInPresentation: JellyPresentation {
                 backgroundStyle: JellyConstants.BackgroundStyle = .dimmed,
                 jellyness: JellyConstants.Jellyness = .none,
                 duration: JellyConstants.Duration = .normal,
-                sizeForViewController: CGSize = CGSize(width:300,height:300),
                 directionShow: JellyConstants.Direction = .top,
-                directionDismiss: JellyConstants.Direction = .top) {
+                directionDismiss: JellyConstants.Direction = .top,
+                widthForViewController: JellyConstants.Size = .fullscreen,
+                heightForViewController: JellyConstants.Size = .fullscreen,
+                horizontalAlignment: JellyConstants.HorizontalAlignment = .center,
+                verticalAlignment: JellyConstants.VerticalAlignment = .center,
+                marginGuards: UIEdgeInsets = .zero) {
         
         self.dismissCurve = dismissCurve
         self.presentationCurve = presentationCurve
@@ -34,8 +44,12 @@ public struct JellySlideInPresentation: JellyPresentation {
         self.backgroundStyle = backgroundStyle
         self.jellyness = jellyness
         self.duration = duration
-        self.sizeForViewController = sizeForViewController
         self.directionShow = directionShow
         self.directionDismiss = directionDismiss
+        self.widthForViewController = widthForViewController
+        self.heightForViewController = heightForViewController
+        self.verticalAlignemt = verticalAlignment
+        self.horizontalAlignment = horizontalAlignment
+        self.marginGuards = marginGuards
     }
 }