+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.2 (2016-11-27)
+
+  - New size options (fullscreen, halfscreen and custom) for width and height instead of plain CGSize
+  - New alignment options for NonFullscreen Presentations
+  - New marging guards - if the size you specified is bigger than the screen, the margin guards kick in and will be applied to your vc
+  - If you want to size your vc using margin you can use .fullscreen as a size and set the marginGuards to constraint your vcs dimensions
+  - Better Code documentation and overall structure
+  - Even more Example inside the Readme and Example App
+
 ## 1.0.1 (2016-11-20)
 
 Minor Version update: